+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.8.0] - 2018-09-07
 ### Added
 - Official support for Python 3.7.
 - Add optional `hostname` parameter to `ctds.connect()`.
@@ -208,7 +210,8 @@ _NVARCHAR_ arguments to remote procedure calls.
 ## [1.0.0] - 2016-03-14
 Initial Release
 
-[Unreleased]: https://github.com/zillow/ctds/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/zillow/ctds/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/zillow/ctds/compare/v1.8.0...v1.7.0
 [1.7.0]: https://github.com/zillow/ctds/compare/v1.7.0...v1.6.3
 [1.6.3]: https://github.com/zillow/ctds/compare/v1.6.3...v1.6.2
 [1.6.2]: https://github.com/zillow/ctds/compare/v1.6.2...v1.6.1